NIFI-9701 - Updated test files to have last modified time of epoch to avoid intermittent issue with MIN_AGE property. Moved MockCacheService to its own file.
